requirements.txt
blake3==0.2.2
pyblake2==1.1.2
pycryptodome==3.14.0
xxhash==2.0.2
